added Burst-RM and removed it from SEIFDA.py
